fix(openai): de-duplicate v1 streaming reader into shared internal package
